[refactor] Miss와 Spare 분리 및 예외 처리 코드 추가
