+++ src/main/java/bowling/domain/frame/status/Miss.java
@@ -1,5 +1,6 @@
 package bowling.domain.frame.status;
 
+import bowling.bowlingexception.InvalidPinStatusException;
 import bowling.domain.frame.DownedPin;
 
 public class Miss extends Ended {
@@ -7,6 +8,10 @@ public class Miss extends Ended {
     private final DownedPin secondPitch;
 
     public Miss(DownedPin firstPitch, DownedPin secondPitch) {
+        if (firstPitch.isSpare(secondPitch)) {
+            throw new InvalidPinStatusException();
+        }
+
         this.firstPitch = firstPitch;
         this.secondPitch = secondPitch;
     }
